Move Filament Change Menu to its own file
